Add dependencies and move files around, not sure if we want this to rely on nypr-auth
