+++ package.json
@@ -24,7 +24,6 @@
   },
   "devDependencies": {
     "broccoli-asset-rev": "^2.4.5",
-    "ember-hifi": "^1.10.0",
     "ember-ajax": "^3.0.0",
     "ember-cli": "2.13.0",
     "ember-cli-dependency-checker": "^1.3.0",
@@ -37,11 +36,12 @@
     "ember-cli-uglify": "^1.2.0",
     "ember-disable-prototype-extensions": "^1.1.0",
     "ember-export-application-global": "^2.0.0",
+    "ember-get-config": "^0.2.2",
+    "ember-hifi": "^1.10.0",
     "ember-load-initializers": "^1.0.0",
     "ember-metrics": "0.10.0",
     "ember-poll": "^1.2.0",
     "ember-resolver": "^4.0.0",
-    "ember-metrics": "0.10.0",
     "ember-sinon": "0.7.0",
     "ember-sinon-qunit": "^1.6.0",
     "ember-source": "~2.13.0",
